+++ app/js/main.js
@@ -1,6 +1,9 @@
 // Modules to control application life and create native browser window
 const { app, BrowserWindow, Menu, ipcMain, webContents } = require('electron')
 const { join } = require('path')
+const Store = require('electron-store')
+
+Store.initRenderer()
 
 let mainWindow
 
@@ -46,11 +49,16 @@ function createWindow() {
     width: 1300,
     height: 800,
     webPreferences: {
-      nodeIntegration: true,
+      nodeIntegration: false,
+      contextIsolation: true,
+      enableRemoteModule: false,
+      worldSafeExecuteJavaScript: true,
+      preload: join(__dirname, 'preload.js'),
     }
   })
 
-  Menu.setApplicationMenu(null)
+  //Menu.setApplicationMenu(null)
+  mainWindow.webContents.openDevTools();
 
   mainWindow.loadFile(join(__dirname, '../index.html'))
 
@@ -60,19 +68,25 @@ function createWindow() {
 // This method will be called when Electron has finished
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
-app.whenReady().then(createWindow)
+app.whenReady().then(() => {
+  createWindow()
+  
+  app.on('activate', function () {
+    // On macOS it's common to re-create a window in the app when the
+    // dock icon is clicked and there are no other windows open.
+    if (BrowserWindow.getAllWindows().length === 0) {
+      createWindow()
+    }
+  })
+})
 
 // Quit when all windows are closed.
 app.on('window-all-closed', () => {
   // On macOS it is common for applications and their menu bar
   // to stay active until the user quits explicitly with Cmd + Q
-  if (process.platform !== 'darwin') app.quit()
-})
-
-app.on('activate', () => {
-  // On macOS it's common to re-create a window in the app when the
-  // dock icon is clicked and there are no other windows open.
-  if (BrowserWindow.getAllWindows().length === 0) createWindow()
+  if (process.platform !== 'darwin') {
+    app.quit()
+  }
 })
 
 ipcMain.on('fetch-playlists', (event) => {